requirements.txt
numpy
Pillow
pathlib
